cts/cts_pc_tests_rf/requirements.txt
altgraph==0.17.4
cachetools==5.3.2
colorama==0.4.6
exceptiongroup==1.1.3
iniconfig==2.0.0
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pluggy==1.3.0
py2exe==0.13.0.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyserial==3.5
PyVISA==1.14.0
pywin32-ctypes==0.2.2
tenma-serial==1.1.1
tomli==2.0.1
typing_extensions==4.8.0
